separate classes in BO namespace
