+++ ValheimBackupShared/BO/Backup.cs
@@ -1,11 +1,6 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using ValheimBackup.FTP;
 using Newtonsoft.Json;
 
 namespace ValheimBackup.BO
