+++ ValheimBackupShared/BO/BackupSettings.cs
@@ -2,10 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using ValheimBackupShared.Properties;
 
 namespace ValheimBackup.BO
@@ -176,188 +173,5 @@ namespace ValheimBackup.BO
         }
 
         //TODO: add naming convention for backup files
-    }
-
-    /// <summary>
-    /// Represents a schedule for backing up world files.
-    /// </summary>
-    [JsonObject(MemberSerialization.OptIn)]
-    public class BackupSchedule : INotifyPropertyChanged
-    {
-        private DateTime? _startDate;
-        private DateTime? _endDate;
-
-
-        [JsonProperty]
-        public DateTime? StartDate
-        {
-            get => _startDate;
-            set
-            {
-                if (_startDate != value)
-                {
-                    _startDate = value;
-                    NotifyPropertyChanged("StartDate");
-                }
-            }
-        }
-
-
-        [JsonProperty]
-        public DateTime? EndDate
-        {
-            get => _endDate;
-            set
-            {
-                if (_endDate != value)
-                {
-                    _endDate = value;
-                    NotifyPropertyChanged("EndDate");
-                }
-            }
-        }
-
-        [JsonProperty]
-        public BackupFrequency Frequency { get; set; }
-
-        public BackupSchedule() { }
-
-        public BackupSchedule(BackupFrequency frequency)
-            : this(frequency, DateTime.Now) { }
-
-        public BackupSchedule(BackupFrequency frequency, DateTime startDate)
-            : this(frequency, startDate, DateTime.Now.AddDays(365)) { }
-
-        public BackupSchedule(BackupFrequency frequency, DateTime startDate, DateTime endDate)
-        {
-            this.StartDate = startDate;
-            this.EndDate = endDate;
-            this.Frequency = frequency;
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        private void NotifyPropertyChanged(string propertyName)
-        {
-            if (PropertyChanged != null)
-            {
-                PropertyChanged(this, new PropertyChangedEventArgs(propertyName));
-            }
-        }
-
-        public override string ToString()
-        {
-            return "every " + Frequency.Amount + " " + Frequency.Period.ToString() + " starting on " + StartDate.ToString();
-        }
-    }
-
-    /// <summary>
-    /// Represents the frequency of occurance for a backup event in the form of "every {amount} {unit}".
-    /// eg. every 2 days, every 30 minutes, every 1 week
-    /// </summary>
-    [JsonObject(MemberSerialization.OptIn)]
-    public class BackupFrequency : INotifyPropertyChanged
-    {
-        private int _amount;
-        private BackupPeriod _period;
-
-        [JsonProperty]
-        public int Amount
-        {
-            get => _amount;
-            set
-            {
-                if (_amount != value)
-                {
-                    _amount = value;
-                    NotifyPropertyChanged("Amount");
-                }
-            }
-        }
-
-        [JsonProperty]
-        public BackupPeriod Period
-        {
-            get => _period;
-            set
-            {
-                if (_period != value)
-                {
-                    _period = value;
-                    NotifyPropertyChanged("Period");
-                }
-            }
-        }
-
-        public BackupFrequency(int amount, BackupPeriod period)
-        {
-            this.Amount = amount;
-            this.Period = period;
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        private void NotifyPropertyChanged(string propertyName)
-        {
-            if (PropertyChanged != null)
-            {
-                PropertyChanged(this, new PropertyChangedEventArgs(propertyName));
-            }
-        }
-    }
-
-    /// <summary>
-    /// Represents the frequency of occurance for cleanup events in the form of "every {amount} {unit}".
-    /// eg. every 2 days, every 30 minutes, every 10 copies
-    /// </summary>
-    [JsonObject(MemberSerialization.OptIn)]
-    public class CleanupFrequency : INotifyPropertyChanged
-    {
-        private int _amount;
-        private CleanupPeriod _period;
-
-        [JsonProperty]
-        public int Amount
-        {
-            get => _amount;
-            set
-            {
-                if (_amount != value)
-                {
-                    _amount = value;
-                    NotifyPropertyChanged("Amount");
-                }
-            }
-        }
-
-        [JsonProperty]
-        public CleanupPeriod Period
-        {
-            get => _period;
-            set
-            {
-                if (_period != value)
-                {
-                    _period = value;
-                    NotifyPropertyChanged("Period");
-                }
-            }
-        }
-
-        public CleanupFrequency(int amount, CleanupPeriod period)
-        {
-            this.Amount = amount;
-            this.Period = period;
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        private void NotifyPropertyChanged(string propertyName)
-        {
-            if (PropertyChanged != null)
-            {
-                PropertyChanged(this, new PropertyChangedEventArgs(propertyName));
-            }
-        }
     }
 }
