+++ ValheimBackupShared/BO/Server.cs
@@ -2,9 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ValheimBackup.BO
 {
